// webui/api/models/SftpService_FileEntry.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
/**
 * File entry representation.
 */
export type SftpService_FileEntry = {
    name?: string;
    path?: string;
    directory?: boolean;
    size?: number;
    modifiedTime?: string;
};